+++ src/main/java/com/ness/movie_release_web/controller/PeopleController.java
@@ -9,11 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.CookieValue;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.security.Principal;
@@ -34,7 +30,7 @@ public class PeopleController {
     }
 
     @GetMapping("/{tmdbId}")
-    public String getDetails(@PathVariable("tmdbId") Integer id,
+    public String getDetails(@PathVariable("tmdbId") Long id,
                              @CookieValue(value = "language", defaultValue = "en") Language language,
                              @CookieValue(value = "mode", defaultValue = "movie") Mode mode,
                              Model model,
