+++ src/main/java/com/ness/movie_release_web/model/wrapper/tmdb/credits/TVCastWrapper.java
@@ -25,16 +25,16 @@ public class TVCastWrapper {
     private List<String> originCountries = new ArrayList<>();
 
     @JsonProperty("episode_count")
-    private Integer episodeCount;
+    private Long episodeCount;
 
     @JsonProperty("vote_count")
-    private Integer voteCount;
+    private Long voteCount;
 
     @JsonProperty("vote_average")
-    private Integer voteAverage;
+    private Long voteAverage;
 
     @JsonProperty("genre_ids")
-    private List<Integer> genreIds = new ArrayList<>();
+    private List<Long> genreIds = new ArrayList<>();
 
     @JsonProperty("name")
     private String name;
@@ -48,7 +48,7 @@ public class TVCastWrapper {
     private String posterPath;
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("original_language")
     private String originalLanguage;
