+++ src/main/java/com/ness/movie_release_web/model/wrapper/tmdb/tvSeries/search/TVSearchWrapper.java
@@ -18,10 +18,10 @@ public class TVSearchWrapper {
     private Integer page;
 
     @JsonProperty("total_results")
-    private Integer totalResults;
+    private Long totalResults;
 
     @JsonProperty("total_pages")
-    private Integer totalPages;
+    private Long totalPages;
 
     @JsonProperty("results")
     private List<TVWrapper> results = new ArrayList<>();
