+++ src/main/java/com/ness/movie_release_web/service/tmdb/Cacheable.java
@@ -9,14 +9,14 @@ import java.util.WeakHashMap;
 
 public abstract class Cacheable<T> {
 
-    private Map<Integer, Map<Language, T>> cache = new WeakHashMap<>();
+    private Map<Long, Map<Language, T>> cache = new WeakHashMap<>();
 
-    protected Optional<T> getFromCache(Integer id, Language language){
+    protected Optional<T> getFromCache(Long id, Language language){
         Map<Language, T> languageMap = cache.computeIfAbsent(id, k -> new HashMap<>());
         return Optional.ofNullable(languageMap.get(language));
     }
 
-    protected void putToCache(Integer id, T object, Language language){
+    protected void putToCache(Long id, T object, Language language){
         Map<Language, T> languageMap = cache.computeIfAbsent(id, k -> new HashMap<>());
         languageMap.put(language, object);
     }
